requirements.lock.txt
accelerate==0.16.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.2
cycler==0.12.1
datasets==4.5.0
diffusers==0.16.1
dill==0.4.0
distlib==0.3.8
exceptiongroup==1.3.1
fairscale==0.4.13
filelock==3.20.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.13.4
idna==3.11
image-reward==1.0
importlib_metadata==8.7.1
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.4.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
propcache==0.4.1
psutil==7.2.1
pyarrow==23.0.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
timm==0.6.13
tokenizers==0.13.3
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==4.26.1
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
virtualenv==20.26.2
wcwidth==0.5.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
